Use nanargmin for finding best trial
The objective value for trials can be NaN, which to Spearmint denotes
"this is a very bad region of the parameter space".  However, using
np.argmin on an array (or list) of values with a NaN will return the
index of a NaN value, not the smallest non-NaN value.  You need to use
np.nanargmin for that.